feat: migrate features worker to next api
