+++ apps/api/.env.example
@@ -3,3 +3,7 @@ SECRET="secret"
 EVER_ACCESS_KEY=""
 EVER_ACCESS_SECRET=""
 LENS_IPFS_AUTH_KEY=""
+KV_REST_API_READ_ONLY_TOKEN="="
+KV_REST_API_TOKEN="="
+KV_REST_API_URL=""
+KV_URL=""
