+++ docs/ports.md
@@ -18,4 +18,3 @@ The following ports are exposed by the repository
 | 8093 | Staff Picks worker    | http://localhost:8093 | https://staff-picks.hey.xyz    |
 | 8094 | Live worker           | http://localhost:8094 | https://live.hey.xyz           |
 | 8095 | Sitemap worker        | http://localhost:8095 | https://sitemap.hey.xyz        |
-| 8096 | Features worker       | http://localhost:8096 | https://features.hey.xyz       |
