+++ packages/data/constants.ts
@@ -92,9 +92,6 @@ export const STAFF_PICKS_WORKER_URL = IS_PRODUCTION
 export const LIVE_WORKER_URL = IS_PRODUCTION
   ? 'https://live.hey.xyz'
   : 'http://localhost:8094';
-export const FEATURES_WORKER_URL = IS_PRODUCTION
-  ? 'https://features.hey.xyz'
-  : 'http://localhost:8096';
 
 // Tokens / Keys
 export const OPENSEA_KEY = '8b95f9e6d52b42fe8c19ddea847c0f5d';
