--- packages/workers/features/.dev.vars.example
@@ -1,1 +1,0 @@
-SUPABASE_KEY=""
