--- packages/workers/features/README.md
@@ -1,1 +1,0 @@
-# Features worker
