--- packages/workers/features/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@workers/features",
-  "version": "0.0.0",
-  "private": true,
-  "license": "AGPL-3.0",
-  "scripts": {
-    "dev": "wrangler dev --port 8096",
-    "lint": "eslint . --ext .ts",
-    "lint:fix": "eslint . --fix --ext .ts",
-    "prettier": "prettier --check \"**/*.{js,ts,tsx,md}\"  --cache",
-    "prettier:fix": "prettier --write \"**/*.{js,ts,tsx,md}\"  --cache",
-    "typecheck": "tsc --pretty",
-    "worker:deploy": "wrangler deploy --var RELEASE:\"$(git rev-parse HEAD)\""
-  },
-  "dependencies": {
-    "@hey/data": "workspace:*",
-    "@hey/lib": "workspace:*",
-    "@hey/supabase": "workspace:*",
-    "itty-router": "^4.0.23",
-    "zod": "^3.22.4"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20231025.0",
-    "@hey/config": "workspace:*",
-    "typescript": "^5.2.2",
-    "wrangler": "^3.15.0"
-  }
-}
