--- packages/workers/features/src/types.ts
@@ -1,14 +1,0 @@
-import type { IRequestStrict } from 'itty-router';
-
-export interface Env {
-  RELEASE: string;
-  SUPABASE_KEY: string;
-  FEATURES: KVNamespace;
-  VERIFIED: KVNamespace;
-}
-
-export type WorkerRequest = {
-  req: Request;
-  env: Env;
-  ctx: ExecutionContext;
-} & IRequestStrict;
