--- packages/workers/features/wrangler.toml
@@ -1,17 +1,0 @@
-name = "features"
-main = "src/index.ts"
-compatibility_date = "2023-10-25"
-keep_vars = true
-
-routes = [
-	{ pattern = "features.hey.xyz", custom_domain = true }
-]
-
-kv_namespaces = [
-  { binding = "FEATURES", id = "7f319321a8b44003b99f9d41e1d7e065", preview_id = "7f319321a8b44003b99f9d41e1d7e065" },
-  { binding = "VERIFIED", id = "b73287bbf01249a5b6efbc529593f994", preview_id = "b73287bbf01249a5b6efbc529593f994" }
-]
-
-[env.production.vars]
-RELEASE = ""
-SUPABASE_KEY = ""
